Split Tendermint RPC client in multiple files
