+++ packages/iov-tendermint-rpc/src/rpcclients/index.ts
@@ -1,0 +1,5 @@
+// This folder contains Tendermint-specific RPC clients
+
+export { instanceOfRpcStreamingClient, RpcClient, RpcStreamingClient } from "./rpcclient";
+export { HttpClient, HttpUriClient } from "./httpclient";
+export { WebsocketClient } from "./websocketclient";
