+++ packages/iov-tendermint-rpc/types/rpcclients/index.d.ts
@@ -1,0 +1,3 @@
+export { instanceOfRpcStreamingClient, RpcClient, RpcStreamingClient } from "./rpcclient";
+export { HttpClient, HttpUriClient } from "./httpclient";
+export { WebsocketClient } from "./websocketclient";
